perf: keep warm control reads out of broad runtime imports
