+++ apps/web/app/api/internal/hosted-execution/assistant-configuration/tool/route.ts
@@ -1,6 +1,6 @@
 import {
   parseHostedRuntimeAssistantConfigurationControlRequest,
-} from "@murphai/hosted-execution/parsers";
+} from "@murphai/hosted-execution/assistant-model";
 
 import {
   handleHostedRuntimeAssistantConfigurationTool,
